[consolidate-create-constants] moves more constants, poplar and sycamore.
